Further GDBus integration
GDBusActionGroup appears to be rather problematic now and might even
have to be removed considering the problems arising from the defunct
cache. Knowledge of the actually exported actions is not updated
adequately. In particular, it is neither updated on creation while the
remote is up-and-running, nor is it updated if the remote spawns lateron
which does not fit well with the GDBusProxy policy of handing a
unavailable remote transparently.
This, combined with the fact that GDBusActionGroup does not provide any
API for syncronously synchronizing with the remote, makes it very
difficult to embed it in any reasonable manner without the help of a lot
of a lot of state-checking and signal-handling.