More fixes to work arround GDBusActionGroup
GDBusActionGroup is kept because it's semantically meaningful in that
these actions are not merely internal communications but actually pose
an interface to the outside world and are, in fact, triggered by running
the executable with command-line switches such as "--hide" or "--show".
Three, possibly four distinct problems can be identified with
GDBusActionGroup (DBAG), these are:
1) Actions are not inquired initially. The DBAG does not make any
attempts by itsself to synchronize its data with the remotely available
actions.
2) If the remote appears on the bus later, DBAG takes no notion of it
and does not attempt to synchronize its data.
3) DBAG does never attempt to update, if it ever had taken notion of any
remote actions, as can be seen by those conditionals which check if the
HashTable is empty.
4) It is quite possible, that DBAG doesn't even update its list of
actions when a chance on the remote happens.