CMake-Magic and GSettings
Settings shall be stored through GSettings, despite its need for a
suitable backend. However, it will allow convenient implementation of
storing and loading preferences. In that context, CMakeLists was fully
extended to conver installation, including compile the GSettings-XML and
updating the immodules cache.
As a sideeffect, gboard will henceforth have to be launched with
XDG_DATA_DIRS correctly pointing to the directory within which the
compiled gsettings schema can be found.