Reworked the whole input logic
A complete overhaul which resulted in a completely new system of a
modifier stack and which now works consistently with X.org! Among many
new features, non-sticky modifiers are now actually held down (as in,
they are not released), which allows one to press Shift, followed by j,
to get a J. While everything was done to make it work with X as well as
possible, attention was also paid to remaining flexible and not relying
on X' quirks.