More fixes on Modstack
All non sticky modifiers are now released when a sticky modifier is
pressed. This is mainly to cope with X11 and the associated emitter and
is not intrinsically necessary, but it also appears to be more
intuitive.
If we would not do that, a sticky modifier (which, intuitively, lies
below non-sticky modifiers, which are applied on top of it, normally)
may elevate keys which the non-sticky modifier(s) had currently evelated
to its/their own state. While, again, this is not intrinsically bad, it
does not play well with the X11 emitter unless one severely bloats the
layout to work around it, because X will note that the non-sticky is
still pressed.
This may be reverted in the future (thus marked as TODO in the code)
when the X11 emitter emits state-independently (by keeping track of
modifiers internally).
For now, the bloat and chaos required in the layout in order to prevent
sticky modifiers from elevating keys which a non-sticky modifier (and
thus, at least in X11, actively pressed modifier) has currently
elevated, would be too immense, so this is somewhat of a bargain towards
X11.
Still, I currently see no situation in which the user would like to
retain the non-sticky keys when pressing a sticky key. If such a
situation arises, things written here are to be reconsidered.