Strange Mis-Optimization
Without the newly introduced line, GBoard appears not to receive any
release event (or perhaps it receives the release before the emission?
No way to tell because every attempt to debug it makes the behaviour
disappear). This should be tracked down and fixed asap. Perhaps there
are more issues which appear only through optimization? We should stay
alert.