Finis.
First version seems to be complete. In the distant future, GBoard should
become a hybrid keyboard/handwriting-recognizer which quickly changes
between the keyboard and "blackboard" depending on the currently used
input device (which can be classified as being dedicated for writing,
typing or no particular method in the preferences). For now, the whole
handwriting section was left out. Things left to do include a better
looking keyboard and fixing the IM Module.